fix: rename folders to proper casing in MainPage
