+++ src/components/MainPage/ProjectInfo/ProjectInfo.module.css
