+++ src/components/MainPage/Thanx/Thanx.module.css
