+++ src/components/MainPage/types.ts
